Renaming organization to account
This is a breaking change without an associated database migration, because older versions of SQLite can't rename columns without manually copying data between tables, and I don't want to depend on the end user to have a sufficiently recent version.  Fortunately, this isn't in production yet, and the only affected database can be scrapped.